♻️ [Refactor]: 마이페이지 - 참여 중/지원 현황 스터디 목록 컴포넌트 FSD 기준으로 디렉토리 분리 및 구조 개선
